requirements.txt
awscli==1.32.0
awscli-local==0.20
boto3==1.34.0
botocore==1.34.0
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
cryptography==42.0.8
dill==0.3.6
dnslib==0.9.24
dnspython==2.6.1
docutils==0.16
ecdsa==0.19.0
idna==3.7
iniconfig==2.0.0
jmespath==0.10.0
localstack==3.5.0
localstack-client==2.5
localstack-core==3.5.0
localstack-ext==3.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
plux==1.11.0
psutil==6.0.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.7.2
s3transfer==0.9.0
semver==3.0.2
six==1.16.0
tabulate==0.9.0
tailer==0.4.1
urllib3==1.26.19
